Removed ChatNewServlet. Used /chat/requests (POST) instead.
