--- src/main/java/com/google/codeu/servlets/ChatNewServlet.java
@@ -1,25 +1,0 @@
-package com.google.codeu.servlets;
-
-import java.io.IOException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import com.google.appengine.api.users.UserService;
-import com.google.appengine.api.users.UserServiceFactory;
-
-@WebServlet("/chat/new")
-public class ChatNewServlet extends HttpServlet{
-  @Override
-  public void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws IOException{
-    
-    UserService userService = UserServiceFactory.getUserService();
-    if (!userService.isUserLoggedIn()) {
-      response.sendRedirect("/index.html");
-      return;
-    }
-
-    
-  }
-}
