+++ src/main/webapp/chat.jsp
@@ -12,7 +12,7 @@
       <div class="dropdown">
         <button class="dropbtn">New conversation</button>
         <div class="dropdown-content">
-          <form action="/chat/new" method="POST">
+          <form action="/chat/requests" method="POST">
             <input type="text" name="nickname" placeholder="Nickname">
             <input type="text" name="invitee" placeholder="Invitee email(s)">
             <button>Submit</button>
@@ -26,7 +26,7 @@
           </div>
         </div>
     </div>
-    
+
     <div id="chat-wrapper">
       <div id="conversations">
         <div style="border-bottom: 1px black solid">Conversations</div>
